+++ crates/controller/src/api/signaling/ws_modules/moderation/rabbitmq.rs
@@ -5,12 +5,18 @@
 use serde::{Deserialize, Serialize};
 use types::core::ParticipantId;
 
+use super::KickScope;
+
 /// Control messages sent between controller modules to communicate changes inside a room
 #[derive(Debug, Serialize, Deserialize)]
 #[serde(rename_all = "lowercase")]
 pub enum Message {
     Kicked(ParticipantId),
     Banned(ParticipantId),
+    Debriefed {
+        kick_scope: KickScope,
+        issued_by: ParticipantId,
+    },
     JoinedWaitingRoom(ParticipantId),
     LeftWaitingRoom(ParticipantId),
     WaitingRoomEnableUpdated,